feat.유저,챌린지 기능 수정

+++ src/challenges/challenges.service.ts
@@ -5,19 +5,28 @@ import { Challenge, ChallengeDocument } from './schema/challenge.schema';
 import { CreateChallengeDto } from './dto/create-challenge.dto';
 import { UpdateChallengeDto } from './dto/update-challenge.dto';
 import { UsersService } from 'src/users/users.service';
-import { User } from 'src/users/schema/user.schema';
+import { Image, ImageDocument } from './schema/image.schema';
+import { Post } from 'src/posts/schema/post.schema';
 
 @Injectable()
 export class ChallengesService {
   constructor(@InjectModel(Challenge.name) private challengeModel: Model<ChallengeDocument>,
+  @InjectModel(Image.name) private imageModel: Model<ImageDocument>,
   @Inject(forwardRef(() => UsersService)) private readonly usersService: UsersService) {}
 
-  async create(user: any, createChallengeDto: CreateChallengeDto) {
-    const createChall = new this.challengeModel(createChallengeDto);
+  async create(user: any, createChallengeDto: CreateChallengeDto, file: Express.Multer.File) {
+    let createChall = await new this.challengeModel(createChallengeDto);
     createChall.user = user._id;
 
-    createChall.save();
-    return '챌린지 생성 완료!';
+    let image = new this.imageModel();
+    image.data = file.buffer;
+    image.contentType = file.mimetype;
+
+    createChall.mainImg = image;
+
+    const result = await createChall.save();
+
+    return result;
   }
   
   async findAll() {
@@ -29,6 +38,27 @@ export class ChallengesService {
     orderByDateChallenge.reverse();
 
     return { ongoingChallenge, orderByUsersChallenge, orderByDateChallenge }
+  }
+
+
+  // 신청한 사람
+  async findById(id: string) {
+    const challenge = await this.challengeModel.findById(id);
+
+    return challenge;
+  }
+
+  // 찜한 사람
+  async findByZzimList(id: string) {
+    const challenge = await this.challengeModel.find({ "like_users": `${id}` });
+
+    return challenge;
+  }
+
+  async findList() {
+    const challenge = await this.challengeModel.find({});
+
+    return challenge;
   }
 
   async findByUser(id: string) {
@@ -44,7 +74,7 @@ export class ChallengesService {
     return { challenge, name, count };
   }
 
-  async update(user:any, id: string, updateChallengeDto: UpdateChallengeDto) {
+  async update(user:any, id: string, updateChallengeDto: UpdateChallengeDto, file: Express.Multer.File) {
     const now = new Date();
     const challenge = await this.challengeModel.findById(id);
 
@@ -60,24 +90,28 @@ export class ChallengesService {
       throw new HttpException("챌린지가 시작되었으므로 수정할 수 없습니다.", HttpStatus.BAD_REQUEST);
     }
 
-    /*
-    if(challenge.user.equals(user._id)) { 
+    const userId = user._id;
+    const challUserId = challenge.user;
+
+    if(!userId.equals(challUserId)) { 
       throw new UnauthorizedException("챌린지 생성자만 수정할 수 있습니다.");
     }
-    */
     
-
+    let image = new this.imageModel();
+    image.data = file.buffer;
+    image.contentType = file.mimetype;
+    
     challenge.description = updateChallengeDto.description;
-    challenge.mainImg = updateChallengeDto.mainImg;
+    challenge.mainImg = image;
     challenge.tag = updateChallengeDto.tag;
 
-    challenge.save();
+    const result = await challenge.save();
 
-    return "수정 완료";
+    return result;
   }
 
 
-  async remove(id: string): Promise<string> {
+  async remove(user: any, id: string) {
     const now = new Date();
 
     const challenge = await this.challengeModel.findById({ _id: id});
@@ -91,6 +125,13 @@ export class ChallengesService {
 
     if(challenge.users.length < 0) {
       throw new HttpException("신청한 사람이 존재하므로 삭제할 수 없습니다.", HttpStatus.BAD_REQUEST);
+    }
+
+    const userId = user._id;
+    const challUserId = challenge.user;
+
+    if(!userId.equals(challUserId)) { 
+      throw new UnauthorizedException("챌린지 생성자만 삭제할 수 있습니다.");
     }
 
     const result = await this.challengeModel.deleteOne({ _id: id });
@@ -116,10 +157,15 @@ export class ChallengesService {
       throw new NotFoundException("찾는 챌린지가 존재하지 않습니다. 다시 확인해 주세요.");
     }
 
-    challenge.users.push(user);
-    challenge.save();
 
-    return "신청 완료";
+    if(challenge.users.includes(user._id)) {
+      throw new HttpException("이미 신청한 챌린지 입니다.", HttpStatus.BAD_REQUEST);
+    }
+
+    challenge.users.push(user);
+    const result = await challenge.save();
+
+    return result;
   }
 
   async zzim(user: any, id: string): Promise<string> {
@@ -134,15 +180,14 @@ export class ChallengesService {
       findChall.like_users.push(user);
 
       challenge = findChall;
-      console.log(challenge);
-      challenge.save();
+      await challenge.save();
 
       return "찜 완료";
     }
 
     // 찜이 되어있는 경우
     challenge.like_users = challenge.like_users.filter((ele) => !userId.equals(new Types.ObjectId(`${ele}`)) );
-    challenge.save();
+    await challenge.save();
     
 
     return "찜 취소";
@@ -167,9 +212,24 @@ export class ChallengesService {
     }
 
     challenge.users = challenge.users.filter((ele) => !userId.equals(new Types.ObjectId(`${ele}`)) );
-    challenge.save();
+    await challenge.save();
     
     
    return "챌린지 신청 취소 완료";
+  }
+
+
+  // 인증 추가
+  async postAdd( id: string, createPost) {
+    let challenge = await this.challengeModel.findById(id);
+
+    if(!challenge) {
+      throw new NotFoundException("찾는 챌린지가 존재하지 않습니다. 다시 확인해 주세요.");
+    }
+    // 두번 확인할 필요가 있나?
+
+    challenge.post.push(createPost._id);
+
+    return await challenge.save();;
   }
 }
